Updating PACKET_ZC_PAR_CHANGE_USER
This is a fix and related to issue #8739

+++ src/map/packets.hpp
@@ -913,7 +913,7 @@ struct PACKET_ZC_PAR_CHANGE_USER {
 	int16 packetType;
 	uint32 gid;
 	int16 type;
-	uint16 value;
+	uint32 value;
 } __attribute__((packed));
 DEFINE_PACKET_HEADER(ZC_PAR_CHANGE_USER, 0x1ab)
